requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.1.0
pydantic==2.5.0
python-multipart==0.0.6